updated recipe page styling and removed edit and delete buttons on recipebook cards

+++ public/js/recipes.js
@@ -1,7 +1,7 @@
 //var db = require("../models");
 
 $(document).ready(function () {
-    $.get("/api/posts/", function(data) {
+    $.get("/api/posts/", function (data) {
         $(document).on("click", "button.deleted", handlePostDelete);
         $(document).on("click", "button.edited", handlePostEdit);
 
@@ -12,45 +12,33 @@ $(document).ready(function () {
         }
         var iddd = data[0].id;
 
-
-
-          function handlePostDelete() {
-            function deletePost(id) {
-                $.ajax({
-                  method: "DELETE",
-                  url: "/api/posts/" + id
-                })
-                  .then(function() {
-                    window.location.href = "/recipebook";
-                  });
-              }
-             deletePost(iddd);
-            //console.log("the delete function is being fired");
+        function handlePostDelete() {
+            var conf = confirm("Are you sure you would like to delete this recipe?");
+            if(conf == true) {
+                function deletePost(id) {
+                    $.ajax({
+                        method: "DELETE",
+                        url: "/api/posts/" + id
+                    })
+                        .then(function () {
+                            window.location.href = "/recipebook";
+                        });
+                }
+                deletePost(iddd);
+            }
         }
-
+        var formattedDate = new Date(data[0].updatedAt);
+        formattedDate = moment(formattedDate).format("MMMM Do YYYY, h:mm:ss a");
+        updatedAt.append(formattedDate);
     });
-    
 
 
-    // This function figures out which post we want to delete and then calls
-    // deletePost
-    
 
-    // This function figures out which post we want to edit and takes it to the
-    // Appropriate url
 
-    //     // Retrieve the template data from the HTML (jQuery is used here).  
-    //     var template = $('#recipe').Handlebars();  
-    //     // Compile the template data into a function  
-    //     //var templateScript = Handlebars.compile(template);  
-    //     // Define data in JSON format.  
-    //     var context = {
-    //         "id":   
-    //         "name": "name",  
-    //         "image": "image"  
-    //     };  
-    //     // Pass Data to template script.  
-    //     var html = template(context);  
-    //     // Insert the HTML code into the page  
-    //     $(document.body).append(html);  
+
+
+
+
+
+
 });
